fractracker-complaints-dev/requirements.txt
backcall==0.2.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
decorator==5.1.0
dnspython==2.1.0
email-validator==1.1.3
filelock==3.3.1
Flask==2.0.2
geographiclib==1.52
geopy==2.2.0
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==3.2
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
numpy==1.19.5
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
protobuf==3.19.1
ptyprocess==0.7.0
py3-validate-email==1.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-http-client==3.3.3
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
selenium==3.141.0
sendgrid==6.9.1
six==1.16.0
starkbank-ecdsa==2.0.2
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5
